Move classes; disable Write; make test compile, sorted & pass
